Update checkForecasts.md
"using DataSet Up specifically" - First sentence.  What is this referring to?

The below directions do not make sense.  Step refers to fields but shows feature classes
## Step 1: Validate individual feature classes
There are three required fields that must be added to the geocoded student attribute table in order for SchoolSite Pro to properly identify students to be excluded or optionally included in the enrollment forecasts.:
1. [Study Areas](../createData/createStudyareas.md)
2. [Schools](../createData/createSchools.md)
3. [Students](../createData/createStudents.md)